requirements.txt
gradio==3.27.0
matplotlib==3.7.1
numpy==1.24.1
opencv_python==4.7.0.72
Pillow==9.3.0
pycocotools==2.0.6
segment_anything==1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
